Strip "BIT" from identifiers.
This took quite long to work out but should be fine now: due to some human mistakes in the spec some identifiers don't conform to the convention that this code expects (see workarounds.rs) and this results in extra long names in the best case and duplicate identifiers in the worst. I've decided to just do a prepass on the registry and remove all such mangled identifiers as any codebase building on this library will be starting from the ground up and won't depend on the "wrong" names.