Completely remove vkparse dependency.
I have been unhappy with the library as I feel it is quite unergonomic due to close adherence to the underlying xml schema, in practice many things such as parsing struct and function declarations are easier when done on raw text rather than matching a spaghetti of enums. Subjectively  it also seems to have improved compile times.