Reword formatting indentation mechanism.
Previously indentation was managed by the Separator struct which computed a series of strings from an iterator and then printed them out with a fixed indentation (just a str specified by the programmer) and newline separation. In cases where there are more levels of indentation this becomes unmaintainable and so we now pipe all the text through a simple state machine that increments indentation on braces and indents on every newline.