The old interner is back!
Frankly there has come nothing but pain from the Lasso interner, the need to have the interner at hand
to resolve a string led to so much ugly code that had to be kept up to date by hand. Now we are using a
heavily modified version of the previous interner, now without unescapable data races and with quite a
lot more unsafe. It is not exactly sound and leaks 1 byte of memory, but it is much much easier to use.